Refresh outage recovery snapshot from b25f7cceab12b7ddd900d633353561bbe3ec751d; unfinished work preserved

+++ rust-modules/src/screens/library/identity.rs
@@ -154,11 +154,11 @@ const fn region_base(region: KeyRegion) -> u32 {
 mod tests {
     use super::*;
 
-    fn section(sid: u8) -> LibrarySectionIdentity {
+    fn section(sid: u16) -> LibrarySectionIdentity {
         LibrarySectionIdentity { sid: crate::plex::ServerId::from_raw(sid), key: 1 }
     }
 
-    fn grid(sid: u8, rk: &str) -> LibraryIdentity {
+    fn grid(sid: u16, rk: &str) -> LibraryIdentity {
         LibraryIdentity::Grid {
             section: section(sid),
             sid: crate::plex::ServerId::from_raw(sid),
